move check_for_errors to export.py
