+++ src/everest/detached/jobs/everserver.py
@@ -25,6 +25,7 @@ from ert.run_models.everest_run_model import EverestRunModel
 from everest import export_to_csv, export_with_progress
 from everest.config import EverestConfig
 from everest.detached import ServerStatus, get_opt_status, update_everserver_status
+from everest.export import check_for_errors
 from everest.simulator import JOB_FAILURE
 from everest.strings import (
     EVEREST,
@@ -306,7 +307,8 @@ def main():
         update_everserver_status(config, ServerStatus.exporting_to_csv)
 
         if config.export is not None:
-            err_msgs, export_ecl = config.export.check_for_errors(
+            err_msgs, export_ecl = check_for_errors(
+                config=config.export,
                 optimization_output_path=config.optimization_output_dir,
                 storage_path=config.storage_dir,
                 data_file_path=config.model.data_file,
